slack_app/requirements.txt
tldextract
urlextract
slackclient